refacotr : 최단 경로 알고리즘을 유틸리티화
